+++ app/logicelement/logicoutput.cpp
@@ -1,11 +1,11 @@
 #include "logicoutput.h"
 
-LogicOutput::LogicOutput( size_t inputSz ) : LogicElement( inputSz, inputSz ) {
+LogicOutput::LogicOutput( int inputSz ) : LogicElement( inputSz, inputSz ) {
 
 }
 
 void LogicOutput::_updateLogic( const std::vector< bool > &inputs ) {
-  for( size_t idx = 0; idx < inputs.size( ); ++idx ) {
+  for( int idx = 0; idx < inputs.size( ); ++idx ) {
     setOutputValue( idx, inputs[ idx ] );
   }
 }
